Moved Controller class to new Controller.dart file, commented out eliminateTarget
